fix: fix license style
